refactor: 카드게임 개선
