+++ backend/src/main/java/coffeeshout/minigame/application/CardGameService.java
@@ -3,8 +3,8 @@ package coffeeshout.minigame.application;
 import coffeeshout.global.ui.WebSocketResponse;
 import coffeeshout.minigame.domain.MiniGameType;
 import coffeeshout.minigame.domain.cardgame.CardGame;
-import coffeeshout.minigame.domain.cardgame.round.RoomRoundManager;
-import coffeeshout.minigame.domain.cardgame.round.RoundManagerRegistry;
+import coffeeshout.minigame.domain.round.RoomRoundManager;
+import coffeeshout.minigame.domain.round.RoundManagerRegistry;
 import coffeeshout.minigame.ui.response.MiniGameStartMessage;
 import coffeeshout.minigame.ui.response.MiniGameStateMessage;
 import coffeeshout.room.domain.JoinCode;
