+++ backend/src/main/java/coffeeshout/minigame/domain/cardgame/CardGame.java
@@ -8,8 +8,8 @@ import coffeeshout.minigame.domain.MiniGameType;
 import coffeeshout.minigame.domain.cardgame.card.Card;
 import coffeeshout.minigame.domain.cardgame.card.CardGameDeckGenerator;
 import coffeeshout.minigame.domain.cardgame.card.Deck;
-import coffeeshout.minigame.domain.cardgame.round.RoundPhase;
-import coffeeshout.minigame.domain.cardgame.round.RoundState;
+import coffeeshout.minigame.domain.round.RoundPhase;
+import coffeeshout.minigame.domain.round.RoundState;
 import coffeeshout.room.domain.Playable;
 import coffeeshout.room.domain.player.Player;
 import coffeeshout.room.domain.player.PlayerName;
