--- backend/src/main/java/coffeeshout/minigame/domain/cardgame/CardGameTaskExecutorsV2.java
@@ -1,12 +1,0 @@
-package coffeeshout.minigame.domain.cardgame;
-
-/**
- * @deprecated 이 클래스는 더 이상 사용되지 않습니다.
- * 새로운 RoundManagerRegistry를 사용하세요.
- * 
- * 컴파일 오류 방지를 위해 임시로 유지되며, 향후 버전에서 완전히 제거될 예정입니다.
- */
-@Deprecated
-public class CardGameTaskExecutorsV2 {
-    // 빈 클래스로 유지 (기존 참조 때문에)
-}
