--- backend/src/main/java/coffeeshout/minigame/domain/cardgame/round/RoundManager.java
@@ -1,13 +1,0 @@
-package coffeeshout.minigame.domain.cardgame.round;
-
-/**
- * @deprecated 기존 RoundManager는 더 이상 사용되지 않습니다.
- * 대신 RoundManagerRegistry와 RoomRoundManager를 사용하세요.
- * 
- * 이 클래스는 하위 호환성을 위해 유지되며, 향후 버전에서 제거될 예정입니다.
- */
-@Deprecated
-public class RoundManager {
-    // 기존 구현은 제거하고 빈 클래스로 유지
-    // 컴파일 에러를 방지하기 위함
-}
