--- backend/src/main/java/coffeeshout/minigame/domain/task/CardGameTaskType.java
@@ -1,10 +1,0 @@
-package coffeeshout.minigame.domain.task;
-
-/**
- * @deprecated 이 클래스는 더 이상 사용되지 않습니다.
- * 새로운 RoundPhase와 RoundPhaseHandler를 사용하세요.
- */
-@Deprecated
-public enum CardGameTaskType {
-    // 빈 enum으로 유지 (기존 참조 때문에)
-}
