--- backend/src/main/java/coffeeshout/minigame/domain/task/ChainedTask.java
@@ -1,10 +1,0 @@
-package coffeeshout.minigame.domain.task;
-
-/**
- * @deprecated 이 클래스는 더 이상 사용되지 않습니다.
- * 새로운 RoundManagerRegistry와 RoomRoundManager를 사용하세요.
- */
-@Deprecated
-public class ChainedTask {
-    // 빈 클래스로 유지 (기존 참조 때문에)
-}
